Add files via upload
The t-test for independent samples comparing the mean Sepal.Width between virginica and versicolor species yielded a p-value of 0.001819483. With a significance level of alpha=0.05, we can reject the null hypothesis. This suggests that there is sufficient evidence to conclude that the mean Sepal.Width of virginica species is different from the mean Sepal.Width of versicolor species.